CRUD User & User Role

+++ safe-foodavenue/admin-panel/pages/do_delete_user_role.php
@@ -1,19 +1,12 @@
-<?php 
+<?php
+  require("../php/config.php");
 
-  session_start();
-  require("../php/config.php"); 
-    
-  $sql = "
-    DELETE FROM sfa_role
-    WHERE role_id = '".$_GET["role_id"]."';
-  ";
+  $sql = "DELETE FROM sfa_role WHERE role_id = '". $_GET["role_id"] . "'";
   mysqli_query($con, $sql);
 
-  echo "
-    <script>
-      alert('ลบข้อมูลเรียบร้อย');
-      window.location = './?content=list-user-role';
-    </script>
+  echo "<script>
+  alert('ลบผู้ใช้งานสำเร็จ');
+  window.location.href='./?content=list-user-role';
+  </script>
   ";
-
 ?>
